Refactor: Split unlink-db into own function; and fix echoes.

+++ dob/controller.py
@@ -31,7 +31,7 @@ from . import __arg0name__
 from . import help_strings
 from .copyright import echo_copyright
 from .cmd_config import get_config_path, furnish_config, replenish_config
-from .helpers import dob_in_user_exit, highlight_value
+from .helpers import click_echo, dob_in_user_exit, highlight_value
 from .migrate import upgrade_legacy_database_instructions
 
 # Disable the python_2_unicode_compatible future import warning.
@@ -65,7 +65,9 @@ class Controller(HamsterControl):
 
     @property
     def data_store_exists_at(self):
-        return _('Data store already exists at {}').format(self.config['db_path'])
+        return _(
+            'Data store already exists at {}'
+        ).format(self.config['db_path'])
 
     @property
     def data_store_url(self):
@@ -111,7 +113,7 @@ class Controller(HamsterControl):
             sys.exit(1)
 
         def help_newbie_onboard():
-            click.echo(
+            click_echo(
                 help_strings.NEWBIE_HELP_ONBOARDING.format(
                     legacy_help=upgrade_legacy_database_instructions(self),
                 ),
@@ -125,35 +127,38 @@ class Controller(HamsterControl):
                 oblige_user_create_store()
 
         def oblige_user_create_config():
-            click.echo(help_strings.NEWBIE_HELP_CREATE_CONFIG, err=True)
+            click_echo(help_strings.NEWBIE_HELP_CREATE_CONFIG, err=True)
 
         def oblige_user_create_store():
-            click.echo(help_strings.NEWBIE_HELP_CREATE_STORE, err=True)
+            click_echo(help_strings.NEWBIE_HELP_CREATE_STORE, err=True)
 
         _insist_germinated()
 
     def create_config(self, force):
         exists = os.path.exists(get_config_path())
         if exists and not force:
-            dob_in_user_exit('Config file exists')
+            dob_in_user_exit(_('Config file exists'))
         nark_config, dob_config, file_path = replenish_config()
         self._adjust_log_level(nark_config)
         self.config = nark_config
         self.client_config = dob_config
-        click.echo(
+        click_echo(
             _('Initialized default Dob configuration at {}').format(
                 highlight_value(file_path),
             )
         )
 
     def create_data_store(self, force):
-        skip_standup = self._check_sqlite_store_ready()
+        skip_standup = self.check_sqlite_store_ready()
         if skip_standup:
             if force:
                 self._reset_data_store()
+                unlinked_db = True
             else:
                 dob_in_user_exit(self.data_store_exists_at)
         self._standup_and_version_store()
+        if unlinked_db:
+            self._announce_recreated_store()
 
     def create_config_and_store(self):
         def _create_config_and_store():
@@ -171,7 +176,7 @@ class Controller(HamsterControl):
             if not os.path.exists(cfg_path):
                 self.create_config(force=False)
             else:
-                click.echo(
+                click_echo(
                     _('Configuration already exists at {}').format(
                         highlight_value(cfg_path),
                     )
@@ -182,9 +187,9 @@ class Controller(HamsterControl):
             # like create_config_maybe, which tells the user if the things exists
             # already, because the storage class, SQLAlchemyStore, blindly calls
             # create_all (in create_storage_tables) without checking if db exists.
-            skip_standup = self._check_sqlite_store_ready()
+            skip_standup = self.check_sqlite_store_ready()
             if skip_standup:
-                click.echo(self.data_store_exists_at)
+                click_echo(self.data_store_exists_at)
             else:
                 self._standup_and_version_store()
 
@@ -209,9 +214,9 @@ class Controller(HamsterControl):
             return
         nark_config['sql_log_level'] = 'WARNING'
 
-    def _check_sqlite_store_ready(self):
+    def check_sqlite_store_ready(self):
         if self.config['db_engine'] != 'sqlite':
-            return False
+            return None
         db_path = self.config['db_path']
         if not os.path.isfile(db_path):
             return False
@@ -224,11 +229,25 @@ class Controller(HamsterControl):
                 'FIXME: Reset non-SQLite data store not supported (yet).'
             ))
         else:
-            db_path = self.config['db_path']
-            os.unlink(db_path)
-            click.echo(
-                _('Recreated data store at {}').format(highlight_value(db_path))
-            )
+            self.must_unlink_db_path(force=True)
+
+    def must_unlink_db_path(self, *_args, force):
+        db_path = self.config['db_path']
+        if not os.path.exists(db_path):
+            return
+        if not os.path.isfile(db_path):
+            dob_in_user_exit(_(
+                'Data store exists but is not a file, so not overwriting {}'
+            ).format(db_path))
+        if not force:
+            dob_in_user_exit(self.data_store_exists_at)
+        os.unlink(db_path)
+
+    def _announce_recreated_store(self):
+        click_echo(
+            _('Recreated data store at {}')
+            .format(highlight_value(self.config['db_path']))
+        )
 
     def _standup_and_version_store(self):
         created_fresh = self.standup_store()
@@ -236,7 +255,7 @@ class Controller(HamsterControl):
             verb = _('created')
         else:
             verb = _('already ready')
-        click.echo(
+        click_echo(
             _('Dob database {verb} at {url}').format(
                 verb=verb, url=highlight_value(self.store.get_db_url()),
             )
